separate location api
